move flattenRoutes() fn to docusaurus-utils
